chore(bptree): clear golangci-lint findings under .github/golangci.yml
Runs clean under the repo-canonical lint invocation
(`go run -modfile misc/devdeps/go.mod github.com/golangci/golangci-lint/v2/cmd/golangci-lint run --config .github/golangci.yml ./tm2/pkg/bptree`).

Fixes applied:

- gofmt / goimports: ran on the package to re-align struct-field
  comments that had drifted, and on test files whose imports needed
  reordering after earlier edits.
- errorlint:
  - Twelve `if err == ErrExportDone` sites across test files now use
    `errors.Is(err, ErrExportDone)`; added the `errors` import where
    missing.
  - `export.go`: `e.err != errExportClosed` → `!errors.Is(e.err,
    errExportClosed)`.
  - Four `fmt.Errorf("…: %v", err)` sites switched to `%w` so errors
    propagate wrapped through the test-failure channels.
- gosec G602 (false-positive on `startBuf[0] = PrefixVal` style
  indexed writes to a fixed-size array): switched to composite-literal
  initialisation so the static analyser can verify the index trivially.
- unconvert: dropped the `uint64(prefixLen)` conversion in
  `readLeafNodeV3` — `prefixLen` is already `uint64` from
  `binary.ReadUvarint`.
- predeclared:
  - `persistence_test.go`: local variable named `make` renamed to
    `newTree`.
  - `stress_test.go`: deleted the local `func max(a, b int)` helper
    that shadowed the Go 1.21 built-in of the same name.
- staticcheck:
  - `bptree_test.go` SA9003 (empty branch): turned the documentation-only
    `if` into a real `t.Fatalf` when the shallow-copy-shares-backing
    premise of the test is violated.
  - `iavl_basic_test.go` SA1019 (deprecated `math/rand.Read`): switched
    to an instance-method `src.Read(key)` off a seeded `*rand.Rand`
    so the call is not deprecated and the test remains deterministic.
  - `iavl_iterator_test.go` S1021: merged `var tree *MutableTree` +
    `tree = NewMutableTreeMem()` into a single `:=`.
  - `iavl_tree_test.go` SA4006: replaced three dead `hash, v, err =
    SaveVersion()` captures with `_` for the unused hash, and one dead
    `version, err = LoadVersion(...)` with `_, err = …`.
  - `prune_test.go` SA4006: the initial `v` from the bootstrap
    SaveVersion was overwritten before read; dropped into `_` and
    hoisted `v` to a pre-loop `var`.
- unused: added `//nolint:unused` annotations to
  `(*ImmutableTree).resolveValue` and `(*LeafNode).valueKeyAt` to
  preserve them as parity API surface on their public types rather
  than deleting them.

+++ tm2/pkg/bptree/node.go
@@ -38,14 +38,14 @@ type Node interface {
 // always took `childMu`, which added a full mutex Lock/Unlock to every
 // traversal step even for already-loaded children.
 type InnerNode struct {
-	nodeKey     *NodeKey
-	numKeys     int16
-	childSizes  [B]int64 // leaf count per child subtree; total = sum(childSizes[:numKeys+1])
-	height      int16    // levels above leaf level (parent of leaves = 1)
-	keys        [B - 1][]byte
-	children    [B][]byte    // serialized NodeKey references (12 bytes each), used for persistence
-	childHashes [B]Hash      // hash of each child subtree
-	childNodes  [B]Node      // in-memory child references (nil == unset; read only after childLoaded bit is set)
+	nodeKey       *NodeKey
+	numKeys       int16
+	childSizes    [B]int64 // leaf count per child subtree; total = sum(childSizes[:numKeys+1])
+	height        int16    // levels above leaf level (parent of leaves = 1)
+	keys          [B - 1][]byte
+	children      [B][]byte     // serialized NodeKey references (12 bytes each), used for persistence
+	childHashes   [B]Hash       // hash of each child subtree
+	childNodes    [B]Node       // in-memory child references (nil == unset; read only after childLoaded bit is set)
 	childLoaded   atomic.Uint32 // bitmap: bit i set iff childNodes[i] is populated
 	miniTree      MiniMerkle    // in-memory only, not serialized
 	miniTreeDirty bool          // true when slot contents changed since last Build
@@ -113,8 +113,11 @@ func (n *LeafNode) valueAt(i int, resolver ValueResolver) ([]byte, error) {
 }
 
 // valueKeyAt returns the external valueKey at slot i, or nil if the
-// slot is inline. Used by orphan tracking paths that only care about
-// external slots.
+// slot is inline. Kept alongside valueAt as a parity accessor for
+// orphan-tracking paths that only care about the external-slot case;
+// currently covered by the callers that walk valueKeys directly.
+//
+//nolint:unused // parity accessor retained as part of the LeafNode API surface
 func (n *LeafNode) valueKeyAt(i int) []byte {
 	if n.inlineMask&(uint32(1)<<uint(i)) != 0 {
 		return nil
@@ -122,8 +125,8 @@ func (n *LeafNode) valueKeyAt(i int) []byte {
 	return n.valueKeys[i]
 }
 
-func (n *InnerNode) GetNodeKey() *NodeKey  { return n.nodeKey }
-func (n *LeafNode) GetNodeKey() *NodeKey   { return n.nodeKey }
+func (n *InnerNode) GetNodeKey() *NodeKey   { return n.nodeKey }
+func (n *LeafNode) GetNodeKey() *NodeKey    { return n.nodeKey }
 func (n *InnerNode) SetNodeKey(nk *NodeKey) { n.nodeKey = nk }
 func (n *LeafNode) SetNodeKey(nk *NodeKey)  { n.nodeKey = nk }
 
@@ -244,7 +247,6 @@ func (n *InnerNode) rebuildChildLoaded() {
 	}
 	n.childLoaded.Store(mask)
 }
-
 
 // RebuildMiniMerkle eagerly recomputes the full mini merkle tree from
 // slot-level hashes. For InnerNode, slots are childHashes. For
@@ -820,7 +822,7 @@ func readLeafNodeV3(nk *NodeKey, r *bytes.Reader) (_ *LeafNode, err error) {
 			if err != nil {
 				return nil, fmt.Errorf("reading key %d suffixLen: %w", i, err)
 			}
-			keyLen := uint64(prefixLen) + suffixLen
+			keyLen := prefixLen + suffixLen
 			if keyLen > maxReadBytesLen {
 				return nil, fmt.Errorf("key %d length %d exceeds maximum %d", i, keyLen, maxReadBytesLen)
 			}
